This should be an int64_t

+++ Teak/TeakCExtern.m
@@ -27,7 +27,7 @@ void TeakTrackEvent(const char* actionId, const char* objectTypeId, const char* 
                             andObjectInstanceId:objectInstanceId == NULL ? nil : [NSString stringWithUTF8String:objectInstanceId]];
 }
 
-void TeakIncrementEvent(const char* actionId, const char* objectTypeId, const char* objectInstanceId, int32_t count) {
+void TeakIncrementEvent(const char* actionId, const char* objectTypeId, const char* objectInstanceId, int64_t count) {
   [[Teak sharedInstance] incrementEventWithActionId:actionId == NULL ? nil : [NSString stringWithUTF8String:actionId]
                                     forObjectTypeId:objectTypeId == NULL ? nil : [NSString stringWithUTF8String:objectTypeId]
                                 andObjectInstanceId:objectInstanceId == NULL ? nil : [NSString stringWithUTF8String:objectInstanceId]
